Remove invoice option on payment form
Remove the ability for customers to add their invoice number to payments. This would cause some customers to error and not submit payment. It's unknown if the issue was on paypal side or if the user just didn't add an invoice number and was prevented from making a payment on the website itself due to the invoice number being required.